backend/requirement.txt
fastapi
uvicorn[standard]
firebase-admin
paho-mqtt
python-dotenv
requests
